I can see hands, and the rig/skin seems ok. But the rest... ouch.
I wanted to keep this model as 1 mesh, instead of cutting off the hands and loading those separate. Results are less than stellar... Lower arms bone are not lining up with the wrists, which I think could be a IK setup issue, but above my limitations to understand how to fix.